Fixed spawn COUNT prop and changed MoveNear accepted args type.
Fixed: Spawn items created with random COUNT property.
Changed: MoveNear functions now accept WORD iStep argument instead of int, since x and y coordinates are stored as WORD.

+++ src/graysvr/CItem.cpp
@@ -1414,7 +1414,7 @@ bool CItem::MoveToCheck( const CPointMap & pt, CChar * pCharMover )
 	return true;
 }
 
-bool CItem::MoveNearObj( const CObjBaseTemplate *pObj, int iSteps )
+bool CItem::MoveNearObj( const CObjBaseTemplate *pObj, WORD iSteps )
 {
 	ADDTOCALLSTACK("CItem::MoveNearObj");
 	// Put in the same container as another item.
